+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [0.9.14] - 2026-09-04
+## [0.9.14-beta.1] - 2026-09-04
 
 ### Added